msys not building due to mismatched gpg keys (for now).
Tracing back, it seems one msys maintainer retired and new
maintainers stepped in (new keys). The long method to fix
this requires curl and grabbing more external things, the
short method just needs to have 'image: Visual Studio 2019'
added to .appveyor.yml

Thanks to genisysram there is new .travis.yml tests for the
powerpc. This patch adds further tests for s390x and arm64.

Having seen 'make' problems while experimenting with PR#23
and the new/additional {ppc64le,s390x,arm64} platforms, it
was worth adding extra info for building libuninameslist-fr

+++ nameslist.c
@@ -307,10 +307,10 @@ int uniNamesList_blockCountAlt(unsigned int lang) {
 
 UN_DLL_EXPORT
 long uniNamesList_blockStartAlt(int uniBlock, unsigned int lang) {
-	int c=-1;
+	long c=-1;
 
 	if ( uniNamesList_haveFR(lang) )
-		c=(int)(uniNamesList_blockStartFR(uniBlock));
+		c=(long)(uniNamesList_blockStartFR(uniBlock));
 	if ( c<0 )
 		c=uniNamesList_blockStart(uniBlock);
 	return( c );
@@ -318,10 +318,10 @@ long uniNamesList_blockStartAlt(int uniBlock, unsigned int lang) {
 
 UN_DLL_EXPORT
 long uniNamesList_blockEndAlt(int uniBlock, unsigned int lang) {
-	int c=-1;
+	long c=-1;
 
 	if ( uniNamesList_haveFR(lang) )
-		c=(int)(uniNamesList_blockEndFR(uniBlock));
+		c=(long)(uniNamesList_blockEndFR(uniBlock));
 	if ( c<0 )
 		c=uniNamesList_blockEnd(uniBlock);
 	return( c );
